requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
attrs==24.2.0
audioop-lts==0.2.1
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
yarl==1.18.0
yfinance==0.2.50
mplfinance==0.12.10b0